Added Adapter design pattern
