+++ SwordsAndSandals/Objects/Classes/Player.cs
@@ -9,6 +9,7 @@ using SwordsAndSandals.Objects.Items.Weapons;
 using SwordsAndSandals.Objects.Stats;
 using System;
 using System.Collections.Generic;
+using System.Reflection.Metadata;
 
 namespace SwordsAndSandals.Objects.Classes
 {
@@ -79,41 +80,10 @@ namespace SwordsAndSandals.Objects.Classes
             }
         }
 
-        //public override void Draw(SpriteBatch batch)
-        //{
-        //    //animation.Draw(batch, new Vector2(Position.X, Position.Y - animation.Scale * animation.frameHeight/2), Origin);
-        //    //int numIcons = buttons.Count;
-        //    //float radius = animation.Scale * animation.frameHeight/2;
-        //    //float angleIncrement = MathHelper.TwoPi / numIcons;
-        //    //int index = 0;
-        //    //foreach (var b in buttons.Values)
-        //    //{
-        //    //    float angle = index * angleIncrement;
-        //    //    float xOffset = -(float)Math.Sin(angle) * radius;
-        //    //    float yOffset = -(float)Math.Cos(angle) * radius;
-        //    //    b.Position = new Vector2(Position.X + xOffset, Position.Y - animation.Scale * (animation.frameHeight / 2 - CorrectionY) + yOffset);
-        //    //    b.Draw(batch);
-        //    //    index++;
-        //    //}
-
-        //    //if(Melee != null) Melee.Draw(batch);
-        //    //if(Ranged != null) Ranged.Draw(batch);
-        //    //if(Shield != null) Shield.Draw(batch);
-        //}
-
-        //public override void Update(GameTime gameTime, List<Sprite> sprites)
-        //{
-        //    //foreach (var b in buttons.Values)
-        //    //{
-        //    //    b.Update(gameTime);
-        //    //}
-        //    animation.Update(gameTime);
-        //    Active.Update(gameTime, this, sprites);
-        //    if (Active != abilities["Idle"] && Active.done == true)
-        //    {
-        //        Active = abilities["Idle"];
-        //        Active.Prepare(this);
-        //    }
-        //}
+        public static Player GetNewPlayer(PlayerFactory playerFactory, ContentManager content, Vector2 position, SpriteEffects flip, bool setButtons)
+        {
+            Player player = playerFactory.CreatePlayer(content, position, flip, setButtons);
+            return player;
+        }
     }
 }
